Fix blender prefabs, add sword model - Fixed my Blender to fbx auto export and removed the manually exported fbx - Redid the scaling and dimensions of the floor and wall tiles so they match perfectly on the grid. - Modeled a sword and attached to player. Logic is similar to recettear where the direction the player is moving towards is the one they are / rotated towards
