+++ Assets/FloorTile.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1001 &5383452009521382756
+--- !u!1001 &4959215740739962421
 PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
@@ -8,59 +8,59 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalScale.x
-      value: 630
+      value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalScale.y
-      value: 630
+      value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalPosition.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalPosition.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalRotation.w
       value: 0.7071067
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalRotation.x
       value: -0.7071068
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalRotation.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalRotation.z
       value: -0
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalEulerAnglesHint.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalEulerAnglesHint.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: -8679921383154817045, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: -7511558181221131132, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_Enabled
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 919132149155446097, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - target: {fileID: 919132149155446097, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       propertyPath: m_Name
       value: FloorTile
       objectReference: {fileID: 0}
@@ -68,22 +68,22 @@ PrefabInstance:
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents:
-    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
       insertIndex: -1
-      addedObject: {fileID: 3272887652027479871}
-  m_SourcePrefab: {fileID: 100100000, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
---- !u!1 &5076881440586577973 stripped
+      addedObject: {fileID: 5966922444342024696}
+  m_SourcePrefab: {fileID: 100100000, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
+--- !u!1 &5193712090759166820 stripped
 GameObject:
-  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-  m_PrefabInstance: {fileID: 5383452009521382756}
+  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
+  m_PrefabInstance: {fileID: 4959215740739962421}
   m_PrefabAsset: {fileID: 0}
---- !u!65 &3272887652027479871
+--- !u!65 &5966922444342024696
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5076881440586577973}
+  m_GameObject: {fileID: 5193712090759166820}
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
@@ -96,5 +96,5 @@ BoxCollider:
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
-  m_Size: {x: 0.020672144, y: 0.02067214, z: 0.02067214}
+  m_Size: {x: 10, y: 9.999999, z: 0.19999997}
   m_Center: {x: 0, y: 0, z: 0}
