+++ Assets/Scripts.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 82a2d4fcab2601e4ab795129f6ede9cb
-PrefabImporter:
+guid: f5f064434fdd6454cbd868d573bda14a
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
