+++ Assets/Scripts/Dungeon.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 699fae8bf64227a49af11d0df750be70
+guid: 68d9f4f6a2ecb2b4dae59e0254ccbbb9
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
