+++ Assets/Sword.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 82a2d4fcab2601e4ab795129f6ede9cb
+guid: 8978018c3d867e248a8b823d55a9458e
 PrefabImporter:
   externalObjects: {}
   userData:
