+++ Assets/WallTile.prefab
@@ -1,89 +1,98 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1001 &5383452009521382756
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalScale.x
-      value: 630
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalScale.y
-      value: 630
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -1.03129
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0.13173676
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 5.918517
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -7511558181221131132, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_Enabled
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 919132149155446097, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      propertyPath: m_Name
-      value: WallTile
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents:
-    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 3272887652027479871}
-  m_SourcePrefab: {fileID: 100100000, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
---- !u!1 &5076881440586577973 stripped
+--- !u!1 &1548362885754179068
 GameObject:
-  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 426bfb62ca37eb24c9cad73d17d9a11c, type: 3}
-  m_PrefabInstance: {fileID: 5383452009521382756}
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
---- !u!65 &3272887652027479871
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 4617988644118833485}
+  - component: {fileID: 1688177433435314770}
+  - component: {fileID: 2196318547523993942}
+  - component: {fileID: 7867895257222974066}
+  m_Layer: 0
+  m_Name: WallTile
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &4617988644118833485
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1548362885754179068}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 1, z: 5}
+  m_LocalScale: {x: 1, y: 0.2, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!33 &1688177433435314770
+MeshFilter:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1548362885754179068}
+  m_Mesh: {fileID: 5062246478642505140, guid: 18cba71de8459d14ea83add3269c3030, type: 3}
+--- !u!23 &2196318547523993942
+MeshRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1548362885754179068}
+  m_Enabled: 1
+  m_CastShadows: 1
+  m_ReceiveShadows: 1
+  m_DynamicOccludee: 1
+  m_StaticShadowCaster: 0
+  m_MotionVectors: 1
+  m_LightProbeUsage: 1
+  m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 2
+  m_RayTraceProcedural: 0
+  m_RayTracingAccelStructBuildFlagsOverride: 0
+  m_RayTracingAccelStructBuildFlags: 1
+  m_RenderingLayerMask: 1
+  m_RendererPriority: 0
+  m_Materials:
+  - {fileID: 2100000, guid: d170887a492509a488ddf60cb65034f7, type: 2}
+  m_StaticBatchInfo:
+    firstSubMesh: 0
+    subMeshCount: 0
+  m_StaticBatchRoot: {fileID: 0}
+  m_ProbeAnchor: {fileID: 0}
+  m_LightProbeVolumeOverride: {fileID: 0}
+  m_ScaleInLightmap: 1
+  m_ReceiveGI: 1
+  m_PreserveUVs: 0
+  m_IgnoreNormalsForChartDetection: 0
+  m_ImportantGI: 0
+  m_StitchLightmapSeams: 1
+  m_SelectedEditorRenderState: 3
+  m_MinimumChartSize: 4
+  m_AutoUVMaxDistance: 0.5
+  m_AutoUVMaxAngle: 89
+  m_LightmapParameters: {fileID: 0}
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
+  m_SortingOrder: 0
+  m_AdditionalVertexStreams: {fileID: 0}
+--- !u!65 &7867895257222974066
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5076881440586577973}
+  m_GameObject: {fileID: 1548362885754179068}
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
