+++ Assets/sword1.blend.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 426bfb62ca37eb24c9cad73d17d9a11c
+guid: 2e5a95867ce546d45ad7238f4a340d6d
 ModelImporter:
   serializedVersion: 22200
   internalIDToNameTable: []
