+++ Assets/sword1.blend1.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 82a2d4fcab2601e4ab795129f6ede9cb
-PrefabImporter:
+guid: 829c899359088a64c99eeb59524fce56
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
